IMAGES FOR ADVERTISER ACCOUNT - Owner and Agency For both images uploaded by advertisers (Profile image and image for your detail page): - Autput: Regardless of file format uploaded by users, every image must be converted in ".jpg" format, if original image is gif or png with transparent background, it must be converted in jpg with white background, like it is for property images. ●  	Then copy the same parameter used to do this image resizes also on admin section. Comment :- Done
1. Active and InActive functionality - It must be verified
As already discussed, this functionality should be used by admin to In-Active and Active a property trough the Button (Publiched / InActive).
This functionality should be a completely independent of the "Suspend and Resume" functionality used by advertisers.
So, I believe you can change status of the column "property_approval" in the DB (1= Publiched and 0 = InActive) and do not change the value of the columns "suspention_status".
Comment :- Done
- we have differentiate (Public/inactive) or (suspended/resume) property in view/edit/list property section

2. Property tabs
If you folow the above point 1, then in the tab  InActive should be disply only the properties InActive by admin, not suspended by advertiser.
Note: for properties suspended by advertiser, can be add a red  text like: “Property suspended by admin” ?

Comment :- Inprocess